API - employee Delete updated

+++ backend/employee/views.py
@@ -157,9 +157,11 @@ class EmployeeDelete(APIView):
         id = input
         if Employee.objects.filter(employee_id = id).count() >= 1:
             employee = Employee.objects.get(employee_id = id)
-            doctor = Doctor.objects.get(employee_id = id)
-            user = User.objects.get(member_id = doctor.doctor_id)
-            user.delete()
+            if employee.employee_role == 'Doctor':
+                doctor = Doctor.objects.get(employee_id = id)
+                doctor.delete()
+                user = User.objects.get(member_id = doctor.doctor_id)
+                user.delete()
             employee.delete()
             response_message = ""
             response_code=""
